Added GCC warnings to the platforms scripts using GCC. Fixed many of the warnings. Simplifed the retrieval of the number of threads and thread num in toast_qarray and toast_cov

+++ src/libtoast/test/toast_test_polyfilter.cpp
@@ -31,7 +31,7 @@ TEST_F( polyfilterTest, filter ) {
     signals[1] = signal2.data();
     signals[2] = signal3.data();
 
-    for ( int i=0; i<n; ++i ) {
+    for ( size_t i=0; i<n; ++i ) {
         signal1[i] = 1;
         signal2[i] = i;
         signal3[i] = i*i;
@@ -44,7 +44,7 @@ TEST_F( polyfilterTest, filter ) {
 
     double rms1start = 0, rms2start = 0, rms3start = 0;
 
-    for ( int i=0; i<n; ++i ) {
+    for ( size_t i=0; i<n; ++i ) {
         rms1start += signal1[i]*signal1[i];
         rms2start += signal2[i]*signal2[i];
         rms3start += signal3[i]*signal3[i];
@@ -59,7 +59,7 @@ TEST_F( polyfilterTest, filter ) {
 
     double rms1 = 0, rms2 = 0, rms3 = 0;
 
-    for ( int i=0; i<n; ++i ) {
+    for ( size_t i=0; i<n; ++i ) {
         rms1 += signal1[i]*signal1[i];
         rms2 += signal2[i]*signal2[i];
         rms3 += signal3[i]*signal3[i];
@@ -83,13 +83,13 @@ TEST_F( polyfilterTest, filter_with_flags ) {
     vector<double> signal3(n);
     vector<unsigned char> flags(n, 0);
 
-    for ( int i=0; i<n/20; ++i ) {
+    for ( size_t i=0; i<n/20; ++i ) {
       flags[i] = 1;
       flags[16*n/20+i] = 1;
     }
 
     long ngood = 0;
-    for ( int i=0; i<n; ++i ) {
+    for ( size_t i=0; i<n; ++i ) {
       if ( flags[i] ) continue;
       ngood++;
     }
@@ -100,7 +100,7 @@ TEST_F( polyfilterTest, filter_with_flags ) {
     signals[1] = signal2.data();
     signals[2] = signal3.data();
 
-    for ( int i=0; i<n; ++i ) {
+    for ( size_t i=0; i<n; ++i ) {
         signal1[i] = 1;
         signal2[i] = i;
         signal3[i] = i*i;
@@ -113,7 +113,7 @@ TEST_F( polyfilterTest, filter_with_flags ) {
 
     double rms1start = 0, rms2start = 0, rms3start = 0;
 
-    for ( int i=0; i<n; ++i ) {
+    for ( size_t i=0; i<n; ++i ) {
         if ( flags[i] ) continue;
         rms1start += signal1[i]*signal1[i];
         rms2start += signal2[i]*signal2[i];
@@ -129,7 +129,7 @@ TEST_F( polyfilterTest, filter_with_flags ) {
 
     double rms1 = 0, rms2 = 0, rms3 = 0;
 
-    for ( int i=0; i<n; ++i ) {
+    for ( size_t i=0; i<n; ++i ) {
         if ( flags[i] ) continue;
         rms1 += signal1[i]*signal1[i];
         rms2 += signal2[i]*signal2[i];
